% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svseq.R
\name{export_bed}
\alias{export_bed}
\title{Export Genomic Coordinates as BED File}
\usage{
export_bed(coords, outfile)
}
\arguments{
\item{coords}{A \code{GenomicRanges} object representing genomic coordinates.}

\item{outfile}{A string specifying the output file path for the BED file.}
}
\value{
Invisible \code{NULL}. The BED file is written to the specified path.
}
\description{
This function exports a set of genomic coordinates to a BED file, suitable for downstream analysis or visualization.
}
\examples{
export_bed(coords = example_sv, outfile = "example.bed")
}
