% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svseq.R
\name{export_fasta}
\alias{export_fasta}
\title{Export DNA Sequence as FASTA File}
\usage{
export_fasta(sequence, outfile, fasta_name = NULL)
}
\arguments{
\item{sequence}{A string representing the DNA sequence to export.}

\item{outfile}{A string specifying the output file path for the FASTA file.}

\item{fasta_name}{An optional string to name the sequence in the FASTA file. Defaults to "fusion_sequence" if not provided.}
}
\value{
Invisible \code{NULL}. The FASTA file is written to the specified path.
}
\description{
This function exports a DNA sequence to a FASTA file, enabling further bioinformatics analyses.
}
\examples{
export_fasta(sequence = "AGCTTGCA", outfile = "sequence.fasta", fasta_name = "example_sequence")
}
