% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svseq.R
\name{svseq}
\alias{svseq}
\title{Extract Fusion Sequence or Coordinates}
\usage{
svseq(
  fusion,
  inversion = FALSE,
  ref = "hg38",
  return = c("sequence", "coords")
)
}
\arguments{
\item{fusion}{fusion string (e.g BCR::chr22:23632645:chr9:133729466::ABL1). General form is \code{upstreamGene:upstreamChrom:upstreamPos-downstreamChrom:downstreamPos::downstreamGene}}

\item{ranges}{A set of genomic ranges representing the derivative chromosome}

\item{sv}{genomic breakpoint, start position, end position,}
}
\value{
the genomic DNA fusion sequence from Gene1 start to Gene2 stop coordinates.
}
\description{
Extract Fusion Sequence or Coordinates
}
\examples{
svseq(
  fusion = "BCR::chr22:23632645|chr9:133729466::ABL1",
  inversion = FALSE,  # Is the first gene inverted in its new position?
  ref = "hg38",
  return = "sequence"
)
}
